getReg/requirements.txt
python-dotenv
google-generativeai
groq
Pillow
gradio
requests
